feat: 소수 계산 로직 추가

+++ src/main/java/calculator/Application.java
@@ -15,7 +15,7 @@ public class Application {
         CalculatorInput calculatorInput = InputParser.toCalculatorInput(input);
         Calculator calculator = new Calculator(calculatorInput.numbers());
 
-        int result = calculator.calculate();
+        double result = calculator.calculate();
 
         OutputFormatter.printResult(result);
     }
